decreased debug animations number. refactor screen.ts, to separated prepare screen code
